services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
redis[asyncio]==5.0.4
meilisearch-python-sdk==2.4.0
minio==7.2.7
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-multipart==0.0.9
cryptography==42.0.8
